Refactoring spannerio - Split logical groupings - Refactor writes to support multiprocessor
