+++ sdks/go/pkg/beam/io/spannerio/read_batch_integration_test.go
