+++ sdks/go/pkg/beam/io/spannerio/write_test.go
@@ -17,114 +17,12 @@ package spannerio
 
 import (
 	"cloud.google.com/go/spanner"
-	"cloud.google.com/go/spanner/spannertest"
 	"cloud.google.com/go/spanner/spansql"
 	"context"
-	"github.com/apache/beam/sdks/v2/go/pkg/beam"
-	"github.com/apache/beam/sdks/v2/go/pkg/beam/testing/passert"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/testing/ptest"
 	"google.golang.org/api/iterator"
-	"google.golang.org/api/option"
-	"google.golang.org/grpc"
-	"reflect"
 	"testing"
 )
-
-type TestDto struct {
-	One string `spanner:"One"`
-	Two int64  `spanner:"Two"`
-}
-
-func TestRead(t *testing.T) {
-	testCases := []struct {
-		name          string
-		database      string
-		rows          []TestDto
-		expectedError bool
-	}{
-		{
-			name:     "Successfully read 4 rows",
-			database: "projects/fake-proj/instances/fake-instance/databases/fake-db-4-rows",
-			rows: []TestDto{
-				{
-					One: "one",
-					Two: 1,
-				},
-				{
-					One: "one",
-					Two: 2,
-				},
-				{
-					One: "one",
-					Two: 3,
-				},
-				{
-					One: "one",
-					Two: 4,
-				},
-			},
-		},
-		{
-			name:     "Successfully read 1 rows",
-			database: "projects/fake-proj/instances/fake-instance/databases/fake-db-1-rows",
-			rows: []TestDto{
-				{
-					One: "one",
-					Two: 1,
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			srv, srvCleanup := newServer(t)
-			defer srvCleanup()
-
-			client, cleanup, err := createFakeClient(srv.Addr, testCase.database)
-			if err != nil {
-				t.Fatalf("Unable to create fake client: %v", err)
-			}
-			defer cleanup()
-
-			ddl, err := spansql.ParseDDL("",
-				`CREATE TABLE Test (
-					One STRING(20),
-					Two INT64,
-				) PRIMARY KEY (Two)`)
-			if err != nil {
-				t.Fatalf("Unable to create DDL statement for spanner test: %v", err)
-			}
-
-			err = srv.UpdateDDL(ddl)
-			if err != nil {
-				t.Fatalf("Unable to run DDL into spanner db: %v", err)
-			}
-
-			var mutations []*spanner.Mutation
-			for _, m := range testCase.rows {
-				mutation, err := spanner.InsertStruct("Test", m)
-				if err != nil {
-					t.Fatalf("Unable to create mutation to insert struct: %v", err)
-				}
-
-				mutations = append(mutations, mutation)
-			}
-
-			_, err = client.Apply(context.Background(), mutations)
-			if err != nil {
-				t.Fatalf("Applying mutations: %v", err)
-			}
-
-			p := beam.NewPipeline()
-			s := p.Root()
-			rows := Query(s, &SpannerDatabase{Client: client}, "SELECT * from Test", reflect.TypeOf(TestDto{}))
-
-			passert.Count(s, rows, "", len(testCase.rows))
-			ptest.RunAndValidate(t, p)
-		})
-	}
-}
 
 func TestWrite(t *testing.T) {
 	testCases := []struct {
@@ -214,34 +112,4 @@ func TestWrite(t *testing.T) {
 			}
 		})
 	}
-}
-
-func newServer(t *testing.T) (*spannertest.Server, func()) {
-	srv, err := spannertest.NewServer("localhost:0")
-	if err != nil {
-		t.Fatalf("Starting in-memory fake spanner: %v", err)
-	}
-
-	return srv, func() {
-		srv.Close()
-	}
-}
-
-func createFakeClient(address string, database string) (*spanner.Client, func(), error) {
-	ctx := context.Background()
-
-	conn, err := grpc.DialContext(ctx, address, grpc.WithInsecure())
-	if err != nil {
-		return nil, nil, err
-	}
-
-	client, err := spanner.NewClient(ctx, database, option.WithGRPCConn(conn))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return client, func() {
-		client.Close()
-		conn.Close()
-	}, nil
 }
